+++ trace-forward/src/Trace/Forward/Utils/ForwardSink.hs
@@ -1,17 +1,11 @@
-{-# LANGUAGE DataKinds #-}
-{-# LANGUAGE GADTs #-}
-
 module Trace.Forward.Utils.ForwardSink
   ( ForwardSink (..)
   ) where
 
-import           Control.Concurrent.STM.TBQueue
-import           Control.Concurrent.STM.TVar
+import Control.Concurrent.STM.TBQueue
 
 data ForwardSink lo = ForwardSink
-  { forwardQueue     :: !(TVar (TBQueue lo))
-  , disconnectedSize :: !Word
-  , connectedSize    :: !Word
-  , wasUsed          :: !(TVar Bool)
+  { forwardQueue     :: !(TBQueue lo)
   , overflowCallback :: !([lo] -> IO ())
   }
+