+++ trace-dispatcher/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Revision history for trace-dispatcher
 
+## NEXT -- Nov 2025
+
+* Replaced both `tofDisconnQueueSize` and `tofConnQueueSize` with `tofQueueSize` (See #6361 for details). 
+
 ## 2.10.0 -- July, 2025
 * Forwarding protocol supports connections over TCP socket, in addition to Unix domain sockets.
 * Failure to initialise the `PrometheusSimple` backend is now lenient - i.e., won't result in an exception being propagated.